Local polynomial estimation of the (intra-curve) semivariogram

Creates a svar.bin (binned semivar. + grid parameters) object with the binned intra-curves semivariances (i.e. computes a binned sample variogram).

# S3 method for npf.data
np.svar(
  x,
  h = NULL,
  maxlag = NULL,
  nlags = round(0.55 * x$grid$n),
  minlag = x$grid$lag,
  degree = 1,
  drv = FALSE,
  hat.bin = TRUE,
  ncv = 0,
  ...
)

# S3 method for npf.locpol
np.svar(
  x,
  var,
  h = NULL,
  maxlag = NULL,
  nlags = round(0.55 * x$grid$n),
  minlag = x$grid$lag,
  degree = 1,
  drv = FALSE,
  hat.bin = TRUE,
  ncv = 0,
  ...
)

Arguments

x

object used to select a method.

h

(full) bandwidth matrix (controls the degree of smoothing; only the upper triangular part of h is used).

maxlag

maximum lag. Defaults to (nlags - 1) * minlag.

nlags

number of lags. Defaults to round(0.55 * x$grid$n).

minlag

minimun lag. Defaults to that of the discretization grid.

degree

degree of the local polynomial used. Defaults to 1 (local linear estimation).

drv

logical; if TRUE, the matrix of estimated first derivatives is returned.

hat.bin

logical; if TRUE, the hat matrix of the binned data is returned.

ncv

integer; determines the number of cells leaved out in each dimension. Defaults to 0 (the full data is used) and it is not normally changed by the user in this setting. See "Details" below.

...

further arguments passed to npsp::svariso.

var

(optional) a vector or an object of class npf.var with the estimated (or theoretical) variances.

Value

Returns an object of S3 class npsp::svar.bin.

Details

Only one of the arguments maxlag or nlags must be specified.

See also